seperates export button component from container
