+++ app/containers/CompList.js
@@ -1,7 +1,6 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import CompList from '../components/CompList';
-// import * as CompListActions from '../actions/complist';
 import * as workspaceActions from '../actions/workspace';
 
 function mapStateToProps(state) {
